Corrected Admin's logic. Added AbstractAdminController.
Fixed the display of the Admin Panel. No tests.

+++ src/main/java/com/javarush/golikov/quest/web/AdminQuestDeleteController.java
@@ -1,24 +1,22 @@
 package com.javarush.golikov.quest.web;
 
 import com.javarush.golikov.quest.service.AdminService;
+import com.javarush.golikov.quest.web.admin.AbstractAdminController;
 import jakarta.servlet.annotation.WebServlet;
-import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
 
 @WebServlet("/admin-quest-delete")
-public class AdminQuestDeleteController extends HttpServlet {
-    private AdminService adminService;
-
-    @Override
-    public void init() {
-        adminService = (AdminService) getServletContext().getAttribute("adminService");
-    }
+public class AdminQuestDeleteController extends AbstractAdminController {
 
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws IOException {
+
+        if (checkAdmin(req, resp)) {
+            return;
+        }
 
         adminService.deleteQuest(req.getParameter("id"));
         resp.sendRedirect(req.getContextPath() + "/admin-quests");
